requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.5
importlib-metadata==6.6.0
itsdangerous==2.1.2
jdatetime==4.1.1
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
pydicom==2.3.1
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
